Make the player level up with enough XP reached & mobs give XP

+++ src/zeldiablo/Character.java
@@ -155,7 +155,7 @@ public class Character extends GameObjects {
         return xp;
     }
 
-    public void setXP(int xp) {
+    public void setXP(double xp) {
         this.xp = xp;
     }
